+++ lib/src/main_action_button_theme.dart
@@ -9,12 +9,29 @@ const _constMargin = EdgeInsets.all(7);
 // ----------------------------------------------------------------------
 
 class MainActionButtonTheme {
+  /// size of [MainActionButtonTheme]
   final double size;
+
+  /// icon that displayed when the
+  /// [BottomBarWithSheet] field [isOpened] == true
   final Widget iconOpened;
+
+  /// icon that displayed when the
+  /// [BottomBarWithSheet] field [isOpened] == false
   final Widget icon;
+
+  /// splash color of widget circle
   final Color splash;
+
+  /// background color of widget circle
   final Color color;
+
+  /// side paddings of [MainActionButton]
   final EdgeInsets margin;
+
+  /// This field can set transform location of [MainActionButton]
+  ///
+  /// Check https://github.com/Frezyx/bottom_bar_with_sheet for understand more
   final Matrix4 transform;
 
   MainActionButtonTheme({